backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai>=1.3.4
qdrant-client==1.9.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0